Fixes, copyright, Employee mutable class
