--- examples/employee-app/src/main/docker/Dockerfile
@@ -1,28 +1,0 @@
-#
-# Copyright (c) 2018, 2019 Oracle and/or its affiliates. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-FROM openjdk:11-jre-slim
-
-RUN mkdir /app
-COPY libs /app/libs
-RUN mkdir /public
-COPY public /public
-RUN mkdir /resources
-COPY resources /resources
-COPY classes/logging.properties /app
-COPY ${project.artifactId}.jar /app
-
-CMD ["java", "-jar", "/app/${project.artifactId}.jar"]
