+++ examples/employee-app/src/main/java/io/helidon/service/employee/Main.java
@@ -1,0 +1,111 @@
+/*
+ * Copyright (c) 2018, 2019 Oracle and/or its affiliates. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.helidon.service.employee;
+
+import java.io.IOException;
+import java.nio.file.Paths;
+import java.util.logging.LogManager;
+
+import io.helidon.config.Config;
+import io.helidon.health.HealthSupport;
+import io.helidon.health.checks.HealthChecks;
+import io.helidon.metrics.MetricsSupport;
+import io.helidon.webserver.Routing;
+import io.helidon.webserver.ServerConfiguration;
+import io.helidon.webserver.StaticContentSupport;
+import io.helidon.webserver.WebServer;
+import io.helidon.media.jsonb.server.JsonBindingSupport;
+
+/**
+ * Simple Hello World rest application.
+ */
+public final class Main {
+
+	/**
+	 * Cannot be instantiated.
+	 */
+	private Main() {
+	}
+
+	/**
+	 * Application main entry point.
+	 * 
+	 * @param args command line arguments.
+	 * @throws IOException if there are problems reading logging properties
+	 */
+	public static void main(final String[] args) throws IOException {
+		startServer();
+	}
+
+	/**
+	 * Start the server.
+	 * 
+	 * @return the created {@link WebServer} instance
+	 * @throws IOException if there are problems reading logging properties
+	 */
+	static WebServer startServer() throws IOException {
+
+		// load logging configuration
+		LogManager.getLogManager().readConfiguration(Main.class.getResourceAsStream("/logging.properties"));
+
+		// By default this will pick up application.yaml from the classpath
+		Config config = Config.create();
+
+		// Get webserver config from the "server" section of application.yaml
+		ServerConfiguration serverConfig = ServerConfiguration.create(config.get("server"));
+
+		WebServer server = WebServer.create(serverConfig, createRouting(config));
+
+		// Try to start the server. If successful, print some info and arrange to
+		// print a message at shutdown. If unsuccessful, print the exception.
+		server.start().thenAccept(ws -> {
+			System.out.println("WEB server is up!");
+            System.out.println("Web client at: http://localhost:" + ws.port() 
+                + "/public/index.html");
+			ws.whenShutdown().thenRun(() -> System.out.println("WEB server is DOWN. Good bye!"));
+		}).exceptionally(t -> {
+			System.err.println("Startup failed: " + t.getMessage());
+			t.printStackTrace(System.err);
+			return null;
+		});
+
+		// Server threads are not daemon. No need to block. Just react.
+
+		return server;
+	}
+
+	/**
+	 * Creates new {@link Routing}.
+	 *
+	 * @return routing configured with JSON support, a health check, and a service
+	 * @param config configuration of this server
+	 */
+	private static Routing createRouting(Config config) {
+
+		MetricsSupport metrics = MetricsSupport.create();
+		EmployeeService employeeService = new EmployeeService(config);
+		HealthSupport health = HealthSupport.builder().add(HealthChecks.healthChecks()) // Adds a convenient set of
+																						// checks
+				.build();
+
+		return Routing.builder().register(JsonBindingSupport.create()).register(health) // Health at "/health"
+				.register(metrics) // Metrics at "/metrics"
+				.register("/employees", employeeService)
+				.register("/public", StaticContentSupport.create(Paths.get("public"))).build();
+	}
+
+}
