+++ examples/employee-app/src/main/java/io/helidon/service/employee/ReposFactory.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright (c) 2018, 2019 Oracle and/or its affiliates. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.helidon.service.employee;
+
+public class ReposFactory {
+
+    public static EmployeeRepository getRepo(String driverType) {
+        switch (driverType) {
+            case "Array":
+                return new EmployeeRepositoryImpl();
+            case "Oracle":
+                return new EmployeeRepositoryImplDB();
+            default:
+                System.out.println("No database driver found...");
+                System.exit(1);
+                return null;
+        }
+
+    }
+}
