--- examples/employee-app/src/main/k8s/app.yaml
@@ -1,63 +1,0 @@
-#
-# Copyright (c) 2018 Oracle and/or its affiliates. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-kind: Service
-apiVersion: v1
-metadata:
-  name: ${project.artifactId}
-  labels:
-    app: ${project.artifactId}
-spec:
-  type: NodePort
-  selector:
-    app: ${project.artifactId}
-  ports:
-  - port: 8080
-    targetPort: 8080
-    name: http
----
-kind: Deployment
-apiVersion: extensions/v1beta1
-metadata:
-  name: ${project.artifactId}
-spec:
-  replicas: 1
-  template:
-    metadata:
-      labels:
-        app: ${project.artifactId}
-        version: v1
-    spec:
-      containers:
-      - name: ${project.artifactId}
-        image: ${project.artifactId}
-        imagePullPolicy: IfNotPresent
-        ports:
-        - containerPort: 8080
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: employee-helidon-lb
-  labels:
-    app: ${project.artifactId}
-spec:
-  type: LoadBalancer
-  ports:
-  - port: 80
-      targetPort: 8080
-  selector:
-    app: ${project.artifactId}
