--- examples/employee-app/src/main/resources/DbCreds.properties
@@ -1,7 +1,0 @@
-/* Copyright 2019 Oracle and/or its affiliates. All rights reserved. */
-
-# Oracle DB Credentials
-user=<user-db>
-password=<password-user-db>
-hosturl=<hostname>:<port>/<database_unique_name>.<host_domain_name>
-drivertype=Array
