+++ examples/employee-app/src/main/resources/META-INF/native-image/helidon-example-reflection-config.json
@@ -1,0 +1,1 @@
+[]
