+++ examples/employee-app/src/main/resources/application.yaml
@@ -15,7 +15,10 @@
 #
 
 app:
-  greeting: "Hello"
+  user: <user-db>
+  password: <password-user-db>
+  hosturl: <hostname>:<port>/<database_unique_name>.<host_domain_name>
+  drivertype: Array
 
 server:
   port: 8080
